refactor webvttfunctional
