+++ viewer/src/components/Player.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import WebVttPlayerFunctional from "./WebVttPlayer/WebVttPlayerFunctional";
+import WebVttPlayer from "./WebVttPlayer/WebVttPlayer";
 
 // functional component PlayerReact that uses ReactPlayer
 // TODO: better parameterize video source and VTT source with props (general spec for 3rd party use!). must define spec.
@@ -17,7 +17,7 @@ export default function Player() {
             <div id="control-panel" className="flex flex-row">
                     <button id="show-metadata" className="bg-gray-200 hover:bg-gray-400 dark:bg-gray-800 dark:hover:bg-gray-700 text-gray-800 dark:text-gray-200 py-2 px-4">Show Metadata</button>
             </div>
-            <WebVttPlayerFunctional
+            <WebVttPlayer
                 preload={false}
                 audio={audioUrl}
                 videoUrl={videoUrl}
